move study index to its own file
