+++ antarest/study/storage/rawstudy/model/filesystem/config/model.py
@@ -24,8 +24,8 @@ from antarest.study.business.model.binding_constraint_model import (
     BindingConstraint,
 )
 from antarest.study.business.model.renewable_cluster_model import RenewableCluster
-from antarest.study.business.model.scenario_builder_model import StudyIndex
 from antarest.study.business.model.sts_model import STStorage, STStorageAdditionalConstraint
+from antarest.study.business.model.study_index import StudyIndex
 from antarest.study.business.model.thermal_cluster_model import ThermalCluster
 from antarest.study.model import STUDY_VERSION_8_7, StudyVersionInt
